isolating version dependent logic
