+++ unlock-js/src/errors.js
@@ -1,0 +1,10 @@
+const Errors = {
+  FATAL_MISSING_PROVIDER: 'FATAL_MISSING_PROVIDER',
+  FATAL_NOT_ENABLED_IN_PROVIDER: 'FATAL_NOT_ENABLED_IN_PROVIDER',
+  FAILED_TO_CREATE_LOCK: 'FAILED_TO_CREATE_LOCK',
+  FAILED_TO_PURCHASE_KEY: 'FAILED_TO_PURCHASE_KEY',
+  FAILED_TO_UPDATE_KEY_PRICE: 'FAILED_TO_UPDATE_KEY_PRICE',
+  FAILED_TO_WITHDRAW_FROM_LOCK: 'FAILED_TO_WITHDRAW_FROM_LOCK',
+}
+
+export default Errors
