+++ unlock-js/src/transactionTypes.js
@@ -1,0 +1,9 @@
+const TransactionType = {
+  LOCK_CREATION: 'LOCK_CREATION',
+  KEY_PURCHASE: 'KEY_PURCHASE',
+  WITHDRAWAL: 'WITHDRAWAL',
+  UPDATE_KEY_PRICE: 'UPDATE_KEY_PRICE',
+  FAILED_TO_CREATE_LOCK: 'FAILED_TO_CREATE_LOCK',
+}
+
+export default TransactionType
