+++ unlock-js/src/v0/index.js
@@ -1,0 +1,23 @@
+import createLock from './createLock'
+import updateKeyPrice from './updateKeyPrice'
+import purchaseKey from './purchaseKey'
+import partialWithdrawFromLock from './partialWithdrawFromLock'
+import withdrawFromLock from './withdrawFromLock'
+import getPastLockTransactions from './getPastLockTransactions'
+import getKeysForLockOnPage from './getKeysForLockOnPage'
+import getPastLockCreationsTransactionsForUser from './getPastLockCreationsTransactionsForUser'
+import getLock from './getLock'
+import getKeyByLockForOwner from './getKeyByLockForOwner'
+
+export default {
+  createLock,
+  updateKeyPrice,
+  purchaseKey,
+  partialWithdrawFromLock,
+  withdrawFromLock,
+  getPastLockTransactions,
+  getKeysForLockOnPage,
+  getPastLockCreationsTransactionsForUser,
+  getLock,
+  getKeyByLockForOwner,
+}
